fix: Constants.h overflow

+++ src/include/Constants.h
@@ -5,8 +5,10 @@
 #ifndef CONSTANTS_H
 #define CONSTANTS_H
 
+#include <cstdint>
+
 constexpr int kDefaultBacklog = 10;
 constexpr int kMaxFileNameSize = 4096;
-constexpr int kMaxFileSize = 10e12;
+constexpr uint64_t kMaxFileSize = 10e12;
 
 #endif //CONSTANTS_H
